+++ integration/spark2/src/main/scala/org/apache/spark/sql/execution/command/ShowCarbonPartitionsCommand.scala
@@ -25,11 +25,12 @@ import scala.collection.mutable.ListBuffer
 import org.apache.spark.sql._
 import org.apache.spark.sql.catalyst.expressions.{Attribute, AttributeReference}
 import org.apache.spark.sql.catalyst.TableIdentifier
-import org.apache.spark.sql.execution.RunnableCommand
+import org.apache.spark.sql.hive.CarbonRelation
 import org.apache.spark.sql.types._
 
 import org.apache.carbondata.common.logging.LogServiceFactory
 import org.apache.carbondata.core.metadata.schema.partition.PartitionType
+
 
 
 private[sql] case class ShowCarbonPartitionsCommand(
@@ -42,12 +43,12 @@ private[sql] case class ShowCarbonPartitionsCommand(
       new MetadataBuilder().putString("comment", "partition id").build())(),
     AttributeReference("Name", StringType, nullable = false,
       new MetadataBuilder().putString("comment", "partition name").build())(),
-    AttributeReference("Value(" + columnName + "=)", StringType, nullable = true,
+    AttributeReference("Value", StringType, nullable = true,
       new MetadataBuilder().putString("comment", "partition value").build())()
   )
-  override def run(sqlContext: SQLContext): Seq[Row] = {
-    val relation = CarbonEnv.get.carbonMetastore
-      .lookupRelation1(tableIdentifier)(sqlContext).
+  override def run(sparkSession: SparkSession): Seq[Row] = {
+    val relation = CarbonEnv.getInstance(sparkSession).carbonMetastore
+      .lookupRelation(tableIdentifier)(sparkSession).
       asInstanceOf[CarbonRelation]
     val carbonTable = relation.tableMeta.carbonTable
     var partitionInfo = carbonTable.getPartitionInfo(
@@ -55,39 +56,41 @@ private[sql] case class ShowCarbonPartitionsCommand(
     var partitionType = partitionInfo.getPartitionType
     var result = Seq.newBuilder[Row]
     columnName = partitionInfo.getColumnSchemaList.get(0).getColumnName
-    if (PartitionType.RANGE.equals(partitionType)) {
-      result.+=(RowFactory.create("0", "", "default"))
-      var id = 1
-      // var name = "partition_"
-      var rangeInfo = partitionInfo.getRangeInfo
-      var size = rangeInfo.size() - 1
-      for (index <- 0 to size) {
-        result.+=(RowFactory.create(id.toString(), "", "< " + rangeInfo.get(index)))
-        id += 1
-      }
-    } else if (PartitionType.RANGE_INTERVAL.equals(partitionType)) {
-      result.+=(RowFactory.create("", "", ""))
-    } else if (PartitionType.LIST.equals(partitionType)) {
-      var id = 0
-      var listInfo = partitionInfo.getListInfo
-      var size = listInfo.size() - 1
-      for (index <- 0 to size) {
-        var listStr = ""
-        listInfo.get(index).toArray().foreach { x =>
-          if (listStr.isEmpty()) {
-            listStr = x.toString()
-          } else {
-            listStr += ", " + x.toString()
-          }
+    LOGGER.info("partition column name:" + columnName)
+    partitionType match {
+      case PartitionType.RANGE =>
+        result.+=(RowFactory.create("0", "", "default"))
+        var id = 1
+        var rangeInfo = partitionInfo.getRangeInfo
+        var size = rangeInfo.size() - 1
+        for (index <- 0 to size) {
+          result.+=(RowFactory.create(id.toString(), "", "< " + rangeInfo.get(index)))
+          id += 1
         }
-        result.+=(RowFactory.create(id.toString(), "", listStr))
-        id += 1
-      }
-    } else if (PartitionType.HASH.equals(partitionType)) {
-      var hashNumber = partitionInfo.getNumPartitions
-      result.+=(RowFactory.create("HASH PARTITION", "", hashNumber.toString()))
-    } else {
-      result.+=(RowFactory.create("", "", ""))
+      case PartitionType.RANGE_INTERVAL =>
+        result.+=(RowFactory.create("", "", ""))
+      case PartitionType.LIST =>
+        result.+=(RowFactory.create("0", "", "default"))
+        var id = 1
+        var listInfo = partitionInfo.getListInfo
+        var size = listInfo.size() - 1
+        for (index <- 0 to size) {
+          var listStr = ""
+          listInfo.get(index).toArray().foreach { x =>
+            if (listStr.isEmpty()) {
+              listStr = x.toString()
+            } else {
+              listStr += ", " + x.toString()
+            }
+          }
+          result.+=(RowFactory.create(id.toString(), "", listStr))
+          id += 1
+        }
+      case PartitionType.HASH =>
+        var hashNumber = partitionInfo.getNumPartitions
+        result.+=(RowFactory.create("HASH PARTITION", "", hashNumber.toString()))
+      case others =>
+        result.+=(RowFactory.create("", "", ""))
     }
     result.result()
   }
